DEHOUMON Christelle/Gestion_plaintes/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
fpdf2==2.7.4
Jinja2==3.1.2